refactor: typeScript 도입 (models, routes, controllers)

+++ src/controllers/room.controller.ts
@@ -1,10 +1,25 @@
+import { Request, Response, NextFunction } from "express";
+
+export interface CustomRequest extends Request {
+  token?: { email: string };
+}
+
+interface User {
+  token?: { email: string };
+}
+
 const createError = require("http-errors");
 const { StatusCodes, ReasonPhrases } = require("http-status-codes");
 
-const Room = require("../models/Room");
-const User = require("../models/User");
+import Room from "../models/Room";
 
-exports.createRoom = async (req, res, next) => {
+import User from "../models/User";
+
+export const createRoom = async (
+  req: Request,
+  res: Response,
+  next: NextFunction,
+) => {
   try {
     const { text_id, userId, roomId, text } = req.body;
     const user = await User.findOne({ email: userId });
@@ -44,7 +59,11 @@ exports.createRoom = async (req, res, next) => {
   }
 };
 
-exports.deleteRoom = async (req, res, next) => {
+export const deleteRoom = async (
+  req: CustomRequest,
+  res: Response,
+  next: NextFunction,
+) => {
   const { roomId } = req.params;
 
   const deleteRooms = async () => {
@@ -55,8 +74,12 @@ exports.deleteRoom = async (req, res, next) => {
         return res.status(404).json({ error: "방을 찾을 수 없습니다." });
       }
 
-      if (room.userId !== req.token.email) {
-        return res.status(403).json({ error: "권한이 없습니다." });
+      if (req.token !== undefined) {
+        console.log(req.token.email);
+      }
+
+      if (!req.token || room.userId.toString() !== req.token.email) {
+        return res.status(403).json({ message: "권한이 없습니다." });
       }
 
       await Room.findOneAndRemove({ _id: roomId });
@@ -70,7 +93,11 @@ exports.deleteRoom = async (req, res, next) => {
   deleteRooms();
 };
 
-exports.getroom = async (req, res, next) => {
+export const getroom = async (
+  req: Request,
+  res: Response,
+  next: NextFunction,
+) => {
   const { userId } = req.params;
 
   try {
